add pages, guest routes, user routes, fix spinner
