--- client/src/components/AddDocumentPage.jsx
@@ -1,13 +1,0 @@
-import DocumentForm from './DocumentForm';
-
-const AddDocument = () => {
-  return (
-    <div className="add-page-body">
-      <div className="container p-4 rounded shadow">
-        <DocumentForm />
-      </div>
-    </div>
-  );
-};
-
-export default AddDocument;
