+++ client/src/components/AppRoutes.jsx
@@ -1,39 +1,8 @@
-import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
-
-import AddDocument from '../components/AddDocumentPage.jsx';
-import Map from '../components/Map.jsx';
 import { useUserContext } from '../contexts/UserContext';
-import { HomeScreen } from '../screens/home/HomeScreen.jsx';
-import LoginForm from '../screens/login/LoginScreen.jsx';
+import { GuestRoutes } from './GuestRoutes.jsx';
+import { UserRoutes } from './UserRoutes.jsx';
 
-const AppRoutes = () => {
+export const AppRoutes = () => {
   const { user } = useUserContext();
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route
-          path="/"
-          element={user ? <HomeScreen /> : <Navigate to="/login" />}
-        />
-        <Route
-          path="/login"
-          element={!user ? <LoginForm /> : <Navigate to="/home" />}
-        />
-        <Route
-          path="/submitDocument"
-          element={user ? <AddDocument /> : <Navigate to="/login" />}
-        />
-        <Route
-          path="/home"
-          element={user ? <HomeScreen /> : <Navigate to="/login" />}
-        />
-        <Route
-          path="/map"
-          element={user ? <Map /> : <Navigate to="/login" />}
-        />
-      </Routes>
-    </BrowserRouter>
-  );
+  return user ? <UserRoutes /> : <GuestRoutes />;
 };
-
-export default AppRoutes;
