+++ client/src/components/style.css
@@ -1,10 +1,3 @@
-.custom-input {
-  background-color: #f8f9fa;
-  border-radius: 8px;
-  border: 1px solid #ced4da;
-  padding: 0.5rem;
-  box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.1);
-}
 
 .document-title {
   font-size: 1.8rem; /* Slightly larger for a more commanding presence */
@@ -22,11 +15,6 @@
 .document-title:hover {
   color: #003d80; /* Darker shade on hover for effect */
   transform: translateY(-2px); /* Subtle lift effect on hover */
-}
-
-.custom-input:focus {
-  border-color: #007bff;
-  box-shadow: 0 0 5px rgba(0, 123, 255, 0.5);
 }
 
 .card {
@@ -52,30 +40,6 @@
 .custom-switch .form-check-input:checked {
   background-color: #003366;
   border-color: #003366;
-}
-
-.stakeholder-label {
-  display: inline-flex;
-  align-items: center;
-  padding: 0.5rem 0.75rem;
-  background-color: #063767a9;
-  color: #ffffff;
-  border-radius: 15px;
-  font-weight: bold;
-  font-size: 0.9rem;
-  gap: 0.5rem;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.remove-label-btn {
-  background: none;
-  border: none;
-  color: #ffffff !important;
-  font-size: 1.2rem;
-  cursor: pointer;
-  line-height: 1;
-  display: flex;
-  align-items: center;
 }
 
 .parent-container {
@@ -156,6 +120,11 @@
   opacity: 0.6;
 }
 
+.btn-custom {
+  background-color: #003366;
+  color: white;
+}
+
 .btn-container {
   margin-top: 20px;
   text-align: center;
@@ -217,16 +186,6 @@
   border: none;
 }
 
-.custom-input {
-  background-color: #f8f9fa;
-  border-radius: 8px;
-  padding: 0.5rem;
-}
-
-.btn-custom {
-  background-color: #003366;
-  color: white;
-}
 
 .modal-dialog {
   max-width: 70%;
@@ -312,13 +271,6 @@
   font-size: 1.1rem;
 }
 
-
-.submit-button-container {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 20px;
-}
-
 .add-page-body {
   padding: 5px;
   transform: scale(0.9);
@@ -339,12 +291,13 @@ body {
   align-items: center;
   justify-content: center;
 }
-#map-wrapper {
+
+.map-wrapper {
   width: 80vw;
   height: 80vh;
   position: relative;
 }
-#map-container {
+.map-container {
   width: 100%;
   height: 100%;
   position: relative;
