+++ client/src/pages/addDocument/AddDocument.css
@@ -1,0 +1,57 @@
+.custom-input:focus {
+    border-color: #007bff;
+    box-shadow: 0 0 5px rgba(0, 123, 255, 0.5);
+  }
+
+  .custom-input {
+    background-color: #f8f9fa;
+    border-radius: 8px;
+    padding: 0.5rem;
+  }
+
+  .custom-input {
+    background-color: #f8f9fa;
+    border-radius: 8px;
+    border: 1px solid #ced4da;
+    padding: 0.5rem;
+    box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.1);
+  }
+
+  .btn-custom {
+    background-color: #003366;
+    color: white;
+  }
+
+  .stakeholder-label {
+    display: inline-flex;
+    align-items: center;
+    padding: 0.5rem 0.75rem;
+    background-color: #063767a9;
+    color: #ffffff;
+    border-radius: 15px;
+    font-weight: bold;
+    font-size: 0.9rem;
+    gap: 0.5rem;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  }
+
+
+.remove-label-btn {
+    background: none;
+    border: none;
+    color: #ffffff !important;
+    font-size: 1.2rem;
+    cursor: pointer;
+    line-height: 1;
+    display: flex;
+    align-items: center;
+  }
+
+  .submit-button-container {
+    display: flex;
+    justify-content: flex-end;
+    margin-top: 20px;
+  }
+  
+  
+
