+++ client/src/pages/home/components/NavComponent.jsx
